Small fixes to unsynchronized_pool_resource

+++ runtime-core/memory-resource/unsynchronized_pool_resource.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <array>
+#include <cstddef>
 
 #include "runtime-core/memory-resource/extra-memory-pool.h"
 #include "runtime-core/memory-resource/details/memory_chunk_list.h"
@@ -140,7 +141,7 @@ private:
   extra_memory_pool *extra_memory_head_{nullptr};
   extra_memory_pool extra_memory_tail_{sizeof(extra_memory_pool)};
 
-  static constexpr size_t MAX_CHUNK_BLOCK_SIZE_{16u * 1024u};
+  static constexpr size_t MAX_CHUNK_BLOCK_SIZE_{static_cast<const size_t>(16U * 1024U)};
   std::array<details::memory_chunk_list, details::get_chunk_id(MAX_CHUNK_BLOCK_SIZE_)> free_chunks_;
 };
